% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actsims-package.R
\docType{package}
\name{actsims-package}
\alias{actsims}
\alias{actsims-package}
\title{actsims: inteRact using R6}
\description{
inteRact, but using R6.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/acastroaraujo/actsims}
  \item \url{https://acastroaraujo.github.io/actsims/}
  \item Report bugs at \url{https://github.com/acastroaraujo/actsims/issues}
}

}
\author{
\strong{Maintainer}: Andrés Castro Araújo \email{andres.castro.araujo@duke.edu}

Authors:
\itemize{
  \item Em Maloney
  \item Aidan Combs
}

}
\keyword{internal}
