% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{build_interact}
\alias{build_interact}
\title{Create a new InteRact object}
\usage{
build_interact(dictionary = "usfullsurveyor2015", equation = "us2010")
}
\arguments{
\item{dictionary}{(character) dictionary name.

See available options using \code{\link[actdata:dataset_keys]{actdata::dataset_keys()}}}

\item{equation}{(character) equation name.

See available options using \code{\link[actdata:equations]{actdata::equations}}

Currently only equations that have "\code{group = all}" are implemented.}
}
\value{
An \code{\link{InteRact}} object.
}
\description{
Create a new \code{\link{InteRact}} by specifying dictionary and
equation names available from the \code{actdata} package.
}
