% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Modifier.R
\name{create_modifier}
\alias{create_modifier}
\title{Identity Modifier}
\usage{
create_modifier(
  dictionary = list("usfullsurveyor2015", "all"),
  equations = list("us2010", "all")
)
}
\arguments{
\item{dictionary}{(character) dictionary name and group. The default is set to \code{list("usfullsurveyor2015", "all")}.

See available options using \code{\link[actdata:dataset_keys]{actdata::dataset_keys()}}}

\item{equations}{(character) equations name and group. The default is set to \code{list("usfullsurveyor2015", "all")}.

See available options using \code{\link[actdata:equations]{actdata::equations}}

Currently only equations with "\code{equation_type == traitid}" are valid.}
}
\value{
A function that takes events of type data.frame(M = c(...), I = (...)) and outputs new dictionary entries
}
\description{
Identity Modifier
}
