% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Situation.R
\name{define_situation}
\alias{define_situation}
\title{Create a Definition of the Situation}
\usage{
define_situation(agent1, agent2)
}
\arguments{
\item{agent1}{an \code{InteRactModel} created by the \code{interact()} function}

\item{agent2}{an \code{InteRactModel} created by the \code{interact()} function}
}
\value{
A \code{Situation} R6 object.
}
\description{
Create a Definition of the Situation
}
