% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/99Experimental.R
\name{define_situation_experimental}
\alias{define_situation_experimental}
\title{Experimental Definition of the Situation}
\usage{
define_situation_experimental(...)
}
\arguments{
\item{...}{various \code{InteRactModel} objects created by the \code{interact()},
they must be named "agent1", "agent2", "agent3", and so on.}
}
\value{
A \code{SituationExperimental} R6 object.
}
\description{
Experimental Definition of the Situation
}
