% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deflection-attributes.R
\name{deflection_attributes}
\alias{deflection_attributes}
\alias{get_fundamentals}
\alias{get_transients}
\alias{get_element_wise_deflection}
\alias{get_long_form}
\title{Extract deflection score attributes}
\usage{
get_fundamentals(x)

get_transients(x)

get_element_wise_deflection(x)

get_long_form(x)
}
\arguments{
\item{x}{an "\code{event_deflection}" object created by the \code{\link[=method-deflection]{$deflection()}} method}
}
\value{
A data frame containing either fundamentals, transients, or element-wise deflection scores
}
\description{
Fundamentals, transients, and element-wise deflection.
}
\examples{
act <- interact("indiana2003", list("nc1978", "female"))

grid <- expand.grid(
  A = "girlfriend",
  B = c("help", "retaliate_against", "torment", "interrogate"),
  O = "boyfriend"
)

d <- act$deflection(grid)
d
get_fundamentals(d)
get_transients(d)
get_element_wise_deflection(d)
get_long_form(d)

}
