% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{interact}
\alias{interact}
\title{Set up and interact object.}
\usage{
interact(dictionary, equation)
}
\arguments{
\item{dictionary}{dictionary name.}

\item{equation}{equation name.}
}
\value{
an "InteRact" object.
}
\description{
Set up and interact object.
}
