% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRactModel.R
\name{method-add-equations}
\alias{method-add-equations}
\alias{add_equation}
\title{Add Other Transformation Equation}
\arguments{
\item{type}{one of "emotionid" or "traitid"}

\item{group}{one of "all" or "female" or "male"}
}
\value{
adds the respective transformation equation to the InteRactModel object
}
\description{
The \verb{$add_equations()} method does this and that..
}
\seealso{
Other InteRactModel methods: 
\code{\link{method-closest-terms}},
\code{\link{method-deflection}},
\code{\link{method-fundamentals}},
\code{\link{method-max-confirm}},
\code{\link{method-modify-identity}},
\code{\link{method-optimal-behavior}},
\code{\link{method-reidentify}}
}
\concept{InteRactModel methods}
