% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRactModel.R
\name{method-closest-terms}
\alias{method-closest-terms}
\alias{closest_terms}
\title{Get closest terms to an EPA profile}
\usage{
closest_terms(
  epa,
  component = c("identity", "behavior", "modifier"),
  max_dist = 1
)
}
\arguments{
\item{epa}{a vector or list of epa ratings}

\item{component}{a vector or list of epa ratings}

\item{max_dist}{a positive real number

It also works with one row data frame with \code{e}, \code{p}, and \code{a} columns}
}
\value{
a list of closest terms found in \verb{$dictionary}, sorted by closeness.
}
\description{
The \verb{$closest_terms()} method does this and that..
}
\examples{
act <- interact()

act$closest_terms(c(e = 2, p = 1, a = 0), component = "identity", max_dist = 0.2)
act$closest_terms(c(e = 2, p = 1, a = 0), component = "behavior", max_dist = 0.2)
act$closest_terms(c(e = 2, p = 1, a = 0), component = "modifier", max_dist = 0.2)

## Using `$closest_terms()` on event deflection data frames

d <- act$deflection(list(A = "ceo", B = "kick", O = "ceo"))
d
opt_reidentify <- act$reidentify(d, who = "object")
opt_reidentify
act$closest_terms(opt_reidentify, max_dist = 1)
opt_behavior <- act$optimal_behavior(d, who = "object")
act$closest_terms(opt_behavior, max_dist = 0.5, component = "behavior")

}
\seealso{
Other InteRactModel methods: 
\code{\link{method-add-equation}},
\code{\link{method-characteristic-emotion}},
\code{\link{method-deflection}},
\code{\link{method-fundamentals}},
\code{\link{method-max-confirm}},
\code{\link{method-modify-identity}},
\code{\link{method-optimal-behavior}},
\code{\link{method-reidentify}}
}
\concept{InteRactModel methods}
