% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRactModel.R
\name{method-deflection}
\alias{method-deflection}
\alias{deflection}
\title{Calculate ABO event deflection scores}
\usage{
deflection(events)
}
\arguments{
\item{events}{a data frame with A, B, and O

Each has to exist within the \verb{$dictionary} field}
}
\value{
An "Event deflection" data frame.

This data frame has an \code{event_deflection} S3 class with custom printing that
works seamlessly with the family of \verb{get_*} functions.
}
\description{
The \verb{$deflection()} method calculates the deflection of an ABO event.
}
\examples{
act <- interact()
act$deflection(data.frame(A = "mother", B = "kick", O = "baby"))

grid <- expand.grid(
  A = c("girlfriend", "medic", "bohemian", "dairy_farmer", "daredevil"),
  B = c("help", "retaliate_against", "torment", "interrogate"),
  O = c("toddler", "capitalist", "boyfriend", "prisoner")
)

head(grid, n = 15)

act$deflection(grid)

}
\seealso{
\code{\link[=get_transients]{get_transients()}}, \code{\link[=get_fundamentals]{get_fundamentals()}}, \code{\link[=get_element_wise_deflection]{get_element_wise_deflection()}}, \code{\link[=get_long_form]{get_long_form()}}

Other InteRactModel methods: 
\code{\link{method-add-equation}},
\code{\link{method-characteristic-emotion}},
\code{\link{method-closest-terms}},
\code{\link{method-fundamentals}},
\code{\link{method-max-confirm}},
\code{\link{method-modify-identity}},
\code{\link{method-optimal-behavior}},
\code{\link{method-reidentify}}
}
\concept{InteRactModel methods}
