% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{method-max-confirm}
\alias{method-max-confirm}
\alias{max_confirm}
\title{Identify the behavior that would maximally confirm the identities of actor
in an actor-object pairing}
\arguments{
\item{events}{a data frame with only only a pair of A, B, or O.}
}
\value{
a data frame with the maximally confirming EPA profiles
}
\description{
The \verb{$max_confirm()} method does this and that..
}
\seealso{
Other InteRactModel methods: 
\code{\link{method-closest-terms}},
\code{\link{method-deflection}},
\code{\link{method-fundamentals}},
\code{\link{method-optimal-behavior}},
\code{\link{method-reidentify}}
}
\concept{InteRactModel methods}
