% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRactModel.R
\name{method-modify-identity}
\alias{method-modify-identity}
\alias{modify_identity}
\title{Create EPA profile for modified identities}
\usage{
modify_identity(events)
}
\arguments{
\item{events}{A data frame with \code{M} (modifier) and \code{I} (identity) columns}
}
\value{
A new dictionary rating for the modified identities
}
\description{
The \verb{$modify_identity()} applies the modifier equations (\code{traitid})
to find the EPA profile of a modified identity.
}
\examples{
act <- interact()

act$add_equation(type = "traitid", group = "all")

new_term <- act$modify_identity(list(M = "angry", I = "doctor"))
new_term
new_term$ratings

grid <- data.frame(
  M = c("tired", "taciturn", "angry", "happy"),
  I = c("academic", "academic", "academic", "academic")
)

act$modify_identity(grid)

}
\seealso{
Other InteRactModel methods: 
\code{\link{method-add-equation}},
\code{\link{method-characteristic-emotion}},
\code{\link{method-closest-terms}},
\code{\link{method-deflection}},
\code{\link{method-fundamentals}},
\code{\link{method-max-confirm}},
\code{\link{method-optimal-behavior}},
\code{\link{method-reidentify}}
}
\concept{InteRactModel methods}
