% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{method-optimal-behavior-object}
\alias{method-optimal-behavior-object}
\alias{optimal_behavior_object}
\title{Calculate the Optimal Behavior for the Object following an Event}
\arguments{
\item{x}{an "Event deflection" object created by the \verb{$deflection()} method}
}
\value{
a data frame of EPA profiles for the optimal behavior of the object
}
\description{
The \verb{$optimal_behavior_object()} method does this and that..
}
\seealso{
Other InteRactModel methods: 
\code{\link{method-closest-terms}},
\code{\link{method-deflection}},
\code{\link{method-max-confirm}},
\code{\link{method-optimal-behavior-actor}}
}
\concept{InteRactModel methods}
