% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRactModel.R
\name{method-optimal-behavior}
\alias{method-optimal-behavior}
\alias{optimal_behavior}
\title{Calculate the Optimal Behavior for the Actor or Object following an Event}
\usage{
optimal_behavior(x, who = c("actor", "object"))
}
\arguments{
\item{x}{an "Event deflection" object created by the \verb{$deflection()} method}

\item{who}{(character) either "actor" or "object"}
}
\value{
a data frame of EPA profiles for the optimal behavior
}
\description{
The \verb{$optimal_behavior()} method does this and that..
}
\seealso{
Other InteRactModel methods: 
\code{\link{method-add-equation}},
\code{\link{method-characteristic-emotion}},
\code{\link{method-closest-terms}},
\code{\link{method-deflection}},
\code{\link{method-fundamentals}},
\code{\link{method-max-confirm}},
\code{\link{method-modify-identity}},
\code{\link{method-reidentify}}
}
\concept{InteRactModel methods}
