% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Situation.R
\name{Situation}
\alias{Situation}
\title{Situation Objects}
\description{
A \code{Situation} object is an \link[R6:R6Class]{R6} object created
by the \code{\link[=define_situation]{define_situation()}} function.

The object stores two InteRactModel objects.

It also provides methods for starting situations (\verb{$start}), adding new
interactions (\verb{$new}), calculating optimal behaviors (\verb{$optimal_behavior}),
and reidentification (\verb{$reidentify}).

You can alternatively use the \verb{sttn_*} family of functions for the same purposes.
}
