% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Situation.R
\name{start_situation}
\alias{start_situation}
\title{Create a Definition of the Situation}
\usage{
start_situation(
  init_event = NULL,
  dictionary = list("usfullsurveyor2015", "all"),
  equations = list("us2010", "all")
)
}
\arguments{
\item{init_event}{an ABO, where A is person1 and O is person2}

\item{dictionary}{(character) dictionary name and group. The default is set to \code{list("usfullsurveyor2015", "all")}.

See available options using \code{\link[actdata:dataset_keys]{actdata::dataset_keys()}}}

\item{equations}{(character) equations name and group. The default is set to \code{list("usfullsurveyor2015", "all")}.

See available options using \code{\link[actdata:equations]{actdata::equations}}

Currently only equations with "\code{equation_type == impressionabo}" are valid.}
}
\value{
A \code{Situation} R6 object.
}
\description{
Create a Definition of the Situation
}
