% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Situation.R
\name{sttn_family}
\alias{sttn_family}
\alias{sttn_activate}
\alias{sttn_start}
\alias{sttn_new}
\alias{sttn_extract}
\title{Situation Functions}
\usage{
sttn_activate(x, which)

sttn_start(x, event)

sttn_new(x, event)

sttn_extract(x)
}
\arguments{
\item{x}{an R6 "\code{Situation}" object created by the \code{define_situation()} function.}

\item{which}{(character) either "agent1" or "agent2."}

\item{event}{an ABO event.}
}
\value{
A modified "\code{Situation}" object.
}
\description{
Start, new, activate, and extract.
}
