% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/actsims-package.R
\docType{package}
\name{actsims-package}
\alias{actsims}
\alias{actsims-package}
\title{actsims: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: andrés castro araújo \email{andres.castro.araujo@duke.edu}

}
\keyword{internal}
