% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deflection-attributes.R
\name{deflection_attributes}
\alias{deflection_attributes}
\alias{get_fundamentals}
\alias{get_transients}
\alias{get_element_wise_deflection}
\title{Deflection Score Attributes}
\usage{
get_fundamentals(x)

get_transients(x)

get_element_wise_deflection(x)
}
\arguments{
\item{x}{an "\code{event_deflection}" object created by the \code{\link[=method-deflection]{$deflection()}} method}
}
\value{
A data frame containing either fundamentals, transients, or element-wise deflection scores
}
\description{
Fundamentals, transients, and element-wise deflection.
}
