% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_fundamentals}
\alias{get_fundamentals}
\alias{get_transients}
\alias{get_element_wise_deflection}
\title{Get Deflection Score Attributes}
\usage{
get_fundamentals(x)

get_transients(x)

get_element_wise_deflection(x)
}
\arguments{
\item{x}{a deflection score}
}
\value{
A matrix object containing either fundamentals, transients, or element-wise deflection scores
}
\description{
Fundamentals, transients, and element-wise deflection.
}
