% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRactModel.R
\name{interact}
\alias{interact}
\title{Create a new InteRactModel object}
\usage{
interact(
  dictionary = list("usfullsurveyor2015", "all"),
  equations = list("us2010", "all")
)
}
\arguments{
\item{dictionary}{(character) dictionary name and group. The default is set to \code{list("usfullsurveyor2015", "all")}

See available options using \code{\link[actdata:dataset_keys]{actdata::dataset_keys()}}}

\item{equations}{(character) equations name and group. The default is set to \code{list("usfullsurveyor2015", "all")}

See available options using \code{\link[actdata:equations]{actdata::equations}}

Currently only equations with "\code{equation_type == impressionabo}" are valid.}
}
\value{
An \code{\link{InteRactModel}} object.
}
\description{
Create a new \code{\link{InteRactModel}} object by specifying dictionary and
equation names available from the \code{actdata} package.
}
