% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sims.R
\name{interact}
\alias{interact}
\title{Set up a design}
\usage{
interact(dictionary, equations)
}
\arguments{
\item{dictionary}{dictionary name}

\item{equations}{equation name}
}
\value{
an "InteRact" object
}
\description{
Set up a design
}
