% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{InteRact}
\alias{InteRact}
\title{InteRact Objects}
\description{
An \code{InteRact} object is an \link[R6:R6Class]{R6} object created
by the \code{\link[=build_interact]{build_interact()}} function.

The object stores (1) a dictionary of EPA ratings
or "fundamentals" and (2) an ACT equation used to calculate "transient impressions."

It also provides methods for calculating deflection scores, behaviors, reidentification,
among others.
}
\section{Methods}{
 \code{InteRact} objects have the following associated
methods, many of which have their own (linked) documentation pages:\tabular{ll}{
   \strong{Method} \tab \strong{Description} \cr
   \code{\link[=method-deflection]{$deflection()}} \tab Return an "Event Deflection" data frame. \cr
   \verb{$optimal_behavior()} \tab Return... \cr
   \verb{$reidentify_object()} \tab Return... \cr
   \verb{$closest_terms()} \tab Return... \cr
}
}

