% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{interact}
\alias{interact}
\title{Create a new InteRactModel object}
\usage{
interact(
  dictionary = "usfullsurveyor2015",
  equations = "us2010",
  group = "all"
)
}
\arguments{
\item{dictionary}{(character) dictionary name.

See available options using \code{\link[actdata:dataset_keys]{actdata::dataset_keys()}}

Currently only dictionaries that have "\code{group == all}" are valid.}

\item{equations}{(character) equations name.

See available options using \code{\link[actdata:equations]{actdata::equations}}

Currently only equations with "\code{equation_type == impressionabo}" are valid.}

\item{group}{(character) one of "all" (default), "male", or "female." This specifies the equation type.}
}
\value{
An \code{\link{InteRactModel}} object.
}
\description{
Create a new \code{\link{InteRactModel}} object by specifying dictionary and
equation names available from the \code{actdata} package.
}
