% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRactModel.R
\name{InteRactModel}
\alias{InteRactModel}
\title{InteRactModel objects}
\description{
An \code{InteRactModel} object is an \link[R6:R6Class]{R6} object created
by the \code{\link[=interact]{interact()}} function.

The object stores (1) a \verb{$dictionary} of EPA ratings and (2) a set of \verb{$equations} used to calculate "transient impressions."
}
\section{Methods}{

\code{InteRactModel} objects have the following associated methods, many of which have their own documentation pages:
\itemize{
\item \code{\link[=method-deflection]{$deflection()}}
\item \code{\link[=method-optimal-behavior]{$optimal_behavior()}}
\item \code{\link[=method-reidentify]{$reidentify()}}
\item \code{\link[=method-max-confirm]{$max_confirm()}}
\item \code{\link[=method-closest-terms]{$closest_terms()}}
\item \code{\link[=method-modify-identity]{$modify_identity()}}
\item \code{\link[=method-characteristic-emotion]{$characteristic_emotion()}}
}
}

