% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRactModel.R
\name{method-characteristic-emotion}
\alias{method-characteristic-emotion}
\alias{characteristic_emotion}
\title{Estimate EPA profile of a characteristic emotion}
\usage{
characteristic_emotion(events)
}
\arguments{
\item{events}{A data frame with an \code{I} column (for "identity")}
}
\description{
Estimate EPA profile of a characteristic emotion
}
\examples{
act <- interact()

act$add_equation(type = "emotionid", group = "male")

act$characteristic_emotion(list(I = "brute"))

x <- list(I = c("censor", "grandson", "infant", "australian", "job_seeker"))
act$characteristic_emotion(x)

}
\seealso{
Other InteRactModel methods: 
\code{\link{method-add-equation}},
\code{\link{method-closest-terms}},
\code{\link{method-deflection}},
\code{\link{method-fundamentals}},
\code{\link{method-max-confirm}},
\code{\link{method-modify-identity}},
\code{\link{method-optimal-behavior}},
\code{\link{method-reidentify}}
}
\concept{InteRactModel methods}
