% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRactModel.R
\name{method-closest-terms}
\alias{method-closest-terms}
\alias{closest_terms}
\title{Get Closest Terms to an EPA profile}
\arguments{
\item{epa}{a vector or list of epa ratings

It also works with one row data frame with \code{e}, \code{p}, and \code{a} columns}
}
\value{
a list of closest terms found in \verb{$dictionary}, sorted by closeness.
}
\description{
The \verb{$closest_terms()} method does this and that..
}
\seealso{
Other InteRactModel methods: 
\code{\link{method-add-equations}},
\code{\link{method-deflection}},
\code{\link{method-fundamentals}},
\code{\link{method-max-confirm}},
\code{\link{method-modify-identity}},
\code{\link{method-optimal-behavior}},
\code{\link{method-reidentify}}
}
\concept{InteRactModel methods}
