% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRactModel.R
\name{method-fundamentals}
\alias{method-fundamentals}
\alias{fundamentals}
\title{Lookup fundamentals in dictionary}
\usage{
fundamentals(x)
}
\arguments{
\item{x}{(character) terms contained within the \verb{$dictionary}}
}
\value{
A data frame of EPA profiles
}
\description{
The \verb{$fundamentals()} method does this and that..
}
\seealso{
\code{\link[=get_fundamentals]{get_fundamentals()}}

Other InteRactModel methods: 
\code{\link{method-add-equations}},
\code{\link{method-closest-terms}},
\code{\link{method-deflection}},
\code{\link{method-max-confirm}},
\code{\link{method-modify-identity}},
\code{\link{method-optimal-behavior}},
\code{\link{method-reidentify}}
}
\concept{InteRactModel methods}
