% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init.R
\name{method-max-confirm}
\alias{method-max-confirm}
\alias{max_confirm}
\title{Identify the behavior that would maximally confirm the identities of actor
in an actor-object pairing}
\arguments{
\item{events}{a data frame with only A and O}
}
\value{
a data frame of EPA profiles for the behavior the maximally confirms the
identity of the actor
}
\description{
The \verb{$max_confirm()} method does this and that..
}
\seealso{
Other InteRactModel methods: 
\code{\link{method-closest-terms}},
\code{\link{method-deflection}},
\code{\link{method-optimal-behavior-actor}},
\code{\link{method-optimal-behavior-object}}
}
\concept{InteRactModel methods}
