% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRactModel.R
\name{method-max-confirm}
\alias{method-max-confirm}
\alias{max_confirm}
\title{Identify the behavior that would maximally confirm the identities of actor
in an actor-object pairing}
\usage{
max_confirm(events, solve_for = c("behavior", "actor", "object"))
}
\arguments{
\item{events}{A data frame with only only a pair of A, B, or O.}

\item{solve_for}{"behavior", "actor", or "object"}
}
\value{
A data frame with the maximally confirming EPA profiles
}
\description{
The \verb{$max_confirm(..., solve_for = "behavior")} method identify the behavior that would maximally confirm the identities of actor
in an actor-object pairing.

does this and that..
}
\examples{
\dontrun{
act$max_confirm(
  events = tibble(A = "god", O = "deadbeat"),
  solve_for = "behavior"
)

act$max_confirm(
  events = list(A = "god", B = "kill"),
  solve_for = "object"
)

act$max_confirm(
  events = data.frame(B = "kill", O = "deadbeat"),
  solve_for = "actor"
)
}

}
\seealso{
Other InteRactModel methods: 
\code{\link{method-add-equation}},
\code{\link{method-characteristic-emotion}},
\code{\link{method-closest-terms}},
\code{\link{method-deflection}},
\code{\link{method-fundamentals}},
\code{\link{method-modify-identity}},
\code{\link{method-optimal-behavior}},
\code{\link{method-reidentify}}
}
\concept{InteRactModel methods}
