% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRactModel.R
\name{method-modify-identity}
\alias{method-modify-identity}
\alias{modify_identity}
\title{Modify Identity}
\arguments{
\item{events}{data.frame(M = c(...), I = (...))}
}
\value{
a new dictionary rating for ththe modified identity
}
\description{
The \verb{$modify_identity()} method does this and that..
}
\seealso{
Other InteRactModel methods: 
\code{\link{method-add-equations}},
\code{\link{method-closest-terms}},
\code{\link{method-deflection}},
\code{\link{method-fundamentals}},
\code{\link{method-max-confirm}},
\code{\link{method-optimal-behavior}},
\code{\link{method-reidentify}}
}
\concept{InteRactModel methods}
