% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InteRactModel.R
\name{method-reidentify}
\alias{method-reidentify}
\alias{reidentify}
\title{Reidentification of actor or object, following an event}
\usage{
reidentify(x, who = c("actor", "object"))
}
\arguments{
\item{x}{an "event deflection" object created by the \verb{$deflection()} method}

\item{who}{(character) either "actor" or "object"}
}
\value{
a data frame of EPA profiles for the optimal reidentification
}
\description{
The \verb{$reidentify()} method does this and that..
}
\seealso{
Other InteRactModel methods: 
\code{\link{method-add-equation}},
\code{\link{method-characteristic-emotion}},
\code{\link{method-closest-terms}},
\code{\link{method-deflection}},
\code{\link{method-fundamentals}},
\code{\link{method-max-confirm}},
\code{\link{method-modify-identity}},
\code{\link{method-optimal-behavior}}
}
\concept{InteRactModel methods}
