% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Situation.R
\name{sttn_family}
\alias{sttn_family}
\alias{sttn_activate}
\alias{sttn_start}
\alias{sttn_new}
\alias{sttn_extract}
\title{Situation Functions}
\usage{
sttn_activate(x, which)

sttn_start(x, event)

sttn_new(x, event)

sttn_extract(x)
}
\arguments{
\item{x}{an R6 "\code{Situation}" object created by the \code{define_situation()} function.}

\item{which}{(character) either "agent1" or "agent2."}

\item{event}{an ABO event.}
}
\value{
A modified "\code{Situation}" object.
}
\description{
Start, new, activate, and extract.
}
\examples{

situation <- define_situation(agent1 = interact(), agent2 = interact())

df <- situation |>
  sttn_activate("agent1") |>
  sttn_start(list(A = "employer", B = "hurt", O = "employee")) |>
  sttn_activate("agent2") |>
  sttn_new(list(A = "employee", B = "pull_away_from", O = "employer")) |>
  sttn_activate("agent1") |>
  sttn_new(list(A = "employer", B = "confront", O = "employee")) |>
  sttn_extract()

  df$deflection

  df$transients

}
